requirements.txt
# Core
python-telegram-bot>=20.7
trafilatura>=1.6.0
newspaper3k>=0.2.8

# LLM
openai>=1.6.0

# Storage & Utils
portalocker>=2.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
httpx>=0.25.0
respx>=0.20.0
